+++ app/Http/Controllers/QuizAPIController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Question;
 use App\Models\Quiz;
 use Exception;
 use Illuminate\Http\Request;
@@ -72,35 +71,4 @@ class QuizAPIController extends Controller
             ], 404);
         }
     }
-
-    public function addNewQuestion( Request $request): JsonResponse
-    {
-        $request->validate([
-            'question' => 'required|string|min:1|max:255',
-            'hint' => 'string|max:255',
-            'points' => 'required|integer',
-
-        ]);
-
-        $question = new Question();
-        $question->question = $request->question;
-        $question->hint = $request->hint;
-        $question->points = $request->points;
-        $question->quiz_id = $request->quiz_id;
-
-        try {
-            $result = $question->save();
-        } catch (Exception $e) {
-            $result = false;
-        }
-
-        if ($result === true) {
-            return response()->json(["message" => 'Question created'], 201);
-        } else {
-            return response()->json([
-                "message" => "Question creation failed"
-            ], 500);
-        }
-    }
-
 }
