+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use \App\Http\Controllers\QuizAPIController;
+use \App\Http\Controllers\QuestionAPIController;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
@@ -11,6 +12,6 @@ Route::get('/user', function (Request $request) {
 Route::get('/quizzes', [QuizAPIController::class, 'displayAllQuizzes']);
 Route::get('/quizzes/{id}', [QuizAPIController::class, 'displayQuizByID']);
 Route::post('/quizzes', [QuizAPIController::class, 'addNewQuiz']);
-Route::post('/questions', [QuizAPIController::class, 'addNewQuestion']);
+Route::post('/questions', [QuestionAPIController::class, 'addNewQuestion']);
